remove repeated file and combine them
